Physics/UI: Add more variety of time speed-up
+ J multiplies the current factor by 10, up to a maximum of 1000; K
  divides it by 10, down to a minimum of 1.

+++ include/graphics.hpp
@@ -17,7 +17,7 @@ struct GraphicsStatus {
     unsigned width, height;
     float z_near, z_far;
     float yfov, aspect;
-    float time_speed_up;
+    int time_speed_up;
 
     float luminance;
 };
